ai-service/requirements.txt
flask==3.1.2
flask-cors==6.0.1
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
